Made unit tests
